Extract cursor.h from misc{1,2}.h and memline.h
