+++ src/nvim/buffer.c
@@ -30,6 +30,7 @@
 #include "nvim/vim.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/digraph.h"
 #include "nvim/eval.h"
