+++ src/nvim/edit.c
@@ -16,6 +16,7 @@
 #include "nvim/edit.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/digraph.h"
 #include "nvim/eval.h"
 #include "nvim/ex_docmd.h"
