+++ src/nvim/ex_docmd.c
@@ -16,6 +16,7 @@
 #include "nvim/ex_docmd.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/digraph.h"
 #include "nvim/edit.h"
