+++ src/nvim/farsi.c
@@ -4,6 +4,7 @@
 ///
 
 
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/ex_docmd.h"
 #include "nvim/ex_eval.h"
