+++ src/nvim/fileio.c
@@ -16,6 +16,7 @@
 #include "nvim/fileio.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
