+++ src/nvim/fold.c
@@ -16,6 +16,7 @@
 #include "nvim/vim.h"
 #include "nvim/fold.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/eval.h"
 #include "nvim/ex_docmd.h"
