+++ src/nvim/indent.c
@@ -1,6 +1,7 @@
 #include "nvim/indent.h"
 #include "nvim/eval.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/memline.h"
 #include "nvim/memory.h"
 #include "nvim/misc1.h"
