+++ src/nvim/normal.c
@@ -18,6 +18,7 @@
 #include "nvim/normal.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/digraph.h"
 #include "nvim/edit.h"
