+++ src/nvim/ops.c
@@ -17,6 +17,7 @@
 #include "nvim/ops.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
