+++ src/nvim/option.c
@@ -39,6 +39,7 @@
 #include "nvim/option.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/digraph.h"
 #include "nvim/eval.h"
