+++ src/nvim/quickfix.c
@@ -16,6 +16,7 @@
 #include "nvim/quickfix.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
