+++ src/nvim/screen.c
@@ -93,6 +93,7 @@
 #include "nvim/screen.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
