+++ src/nvim/search.c
@@ -14,6 +14,7 @@
 #include "nvim/vim.h"
 #include "nvim/search.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
