+++ src/nvim/spell.c
@@ -291,6 +291,7 @@
 #include "nvim/spell.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
