+++ src/nvim/tag.c
@@ -16,6 +16,7 @@
 #include "nvim/tag.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
 #include "nvim/ex_cmds.h"
