+++ src/nvim/window.c
@@ -11,6 +11,7 @@
 #include "nvim/window.h"
 #include "nvim/buffer.h"
 #include "nvim/charset.h"
+#include "nvim/cursor.h"
 #include "nvim/diff.h"
 #include "nvim/edit.h"
 #include "nvim/eval.h"
